Struct arch::fdt::FdtWriter

pub struct FdtWriter {
    data: Vec<u8>,
    off_mem_rsvmap: u32,
    off_dt_struct: u32,
    strings: Vec<u8>,
    string_offsets: BTreeMap<CString, u32>,
    node_depth: usize,
    node_ended: bool,
    boot_cpuid_phys: u32,
}

Interface for writing a Flattened Devicetree (FDT) and emitting a Devicetree Blob (DTB).

Example

use arch::fdt::FdtWriter;

let mut fdt = FdtWriter::new(&[]);
let root_node = fdt.begin_node("")?;
fdt.property_string("compatible", "linux,dummy-virt")?;
fdt.property_u32("#address-cells", 0x2)?;
fdt.property_u32("#size-cells", 0x2)?;
let chosen_node = fdt.begin_node("chosen")?;
fdt.property_u32("linux,pci-probe-only", 1)?;
fdt.property_string("bootargs", "panic=-1 console=hvc0 root=/dev/vda")?;
fdt.end_node(chosen_node)?;
fdt.end_node(root_node)?;
let dtb = fdt.finish(0x1000)?;

Fields
data: Vec<u8>off_mem_rsvmap: u32off_dt_struct: u32strings: Vec<u8>string_offsets: BTreeMap<CString, u32>node_depth: usizenode_ended: boolboot_cpuid_phys: u32
Implementations

impl FdtWriter

pub fn new(mem_reservations: &[FdtReserveEntry]) -> Self

Create a new Flattened Devicetree writer instance.

Arguments

mem_reservations - reserved physical memory regions to list in the FDT header.

fn write_mem_rsvmap(&mut self, mem_reservations: &[FdtReserveEntry])

pub fn set_boot_cpuid_phys(&mut self, boot_cpuid_phys: u32)

Set the boot_cpuid_phys field of the devicetree header.

fn pad(&mut self, num_bytes: usize)

fn align(&mut self, alignment: usize)

fn update_u32(&mut self, offset: usize, val: u32)

fn append_u32(&mut self, val: u32)

fn append_u64(&mut self, val: u64)

pub fn begin_node(&mut self, name: &str) -> Result<FdtWriterNode>

Open a new FDT node.

The node must be closed using end_node.

Arguments

name - name of the node; must not contain any NUL bytes.

pub fn end_node(&mut self, node: FdtWriterNode) -> Result<()>

Close a node previously opened with begin_node.

fn intern_string(&mut self, s: CString) -> u32

pub fn property(&mut self, name: &str, val: &[u8]) -> Result<()>

Write a property.

Arguments

name - name of the property; must not contain any NUL bytes. val - value of the property (raw byte array).

pub fn property_null(&mut self, name: &str) -> Result<()>

Write an empty property.

pub fn property_string(&mut self, name: &str, val: &str) -> Result<()>

Write a string property.

pub fn property_string_list(
&mut self,
name: &str,
values: &[&str]
) -> Result<()>

Write a stringlist property.

pub fn property_u32(&mut self, name: &str, val: u32) -> Result<()>

Write a 32-bit unsigned integer property.

pub fn property_u64(&mut self, name: &str, val: u64) -> Result<()>

Write a 64-bit unsigned integer property.

pub fn property_array_u32(&mut self, name: &str, cells: &[u32]) -> Result<()>

Write a property containing an array of 32-bit unsigned integers.

pub fn property_array_u64(&mut self, name: &str, cells: &[u64]) -> Result<()>

Write a property containing an array of 64-bit unsigned integers.

pub fn finish(self, max_size: usize) -> Result<Vec<u8>>

Finish writing the Devicetree Blob (DTB).

Returns the DTB as a vector of bytes, consuming the FdtWriter. The DTB is always padded up to max_size with zeroes, so the returned value will either be exactly max_size bytes long, or an error will be returned if the DTB does not fit in max_size bytes.

Arguments

max_size - Maximum size of the finished DTB in bytes.