Function arch::serial::add_serial_devices

pub fn add_serial_devices(
    protection_type: ProtectionType, 
    io_bus: &Bus, 
    com_evt_1_3: &Event, 
    com_evt_2_4: &Event, 
    serial_parameters: &BTreeMap<(SerialHardware, u8), SerialParameters>, 
    serial_jail: Option<Minijail>
) -> Result<(), DeviceRegistrationError>

Adds serial devices to the provided bus based on the serial parameters given.

Only devices with hardware type SerialHardware::Serial are added by this function.

Arguments

protection_type - VM protection mode.

io_bus - Bus to add the devices to

com_evt_1_3 - event for com1 and com3

com_evt_1_4 - event for com2 and com4

serial_parameters - definitions of serial parameter configurations.

serial_jail - minijail object cloned for use with each serial device. All four of the traditional PC-style serial ports (COM1-COM4) must be specified.