Function arch::serial::set_default_serial_parameters

pub fn set_default_serial_parameters(
    serial_parameters: &mut BTreeMap<(SerialHardware, u8), SerialParameters>
)

Add the default serial parameters for serial ports that have not already been specified.

This ensures that serial_parameters will contain parameters for each of the four PC-style serial ports (COM1-COM4).

It also sets the first SerialHardware::Serial to be the default console device if no other serial parameters exist with console=true and the first serial device has not already been configured explicitly.