Trait arch::LinuxArch

pub trait LinuxArch {
    type Error: StdError;
    fn guest_memory_layout(
        components: &VmComponents
    ) -> Result<Vec<(GuestAddress, u64)>, Self::Error>;

    fn get_system_allocator_config<V: Vm>(vm: &V) -> SystemAllocatorConfig;

    fn build_vm<V, Vcpu>(
        components: VmComponents, 
        vm_evt_wrtube: &SendTube, 
        system_allocator: &mut SystemAllocator, 
        serial_parameters: &BTreeMap<(SerialHardware, u8), SerialParameters>, 
        serial_jail: Option<Minijail>, 
        battery: (&Option<BatteryType>, Option<Minijail>), 
        vm: V, 
        ramoops_region: Option<RamoopsRegion>, 
        devices: Vec<(Box<dyn BusDeviceObj>, Option<Minijail>)>, 
        irq_chip: &mut dyn IrqChipArch, 
        vcpu_ids: &mut Vec<usize>, 
        debugcon_jail: Option<Minijail>
    ) -> Result<RunnableLinuxVm<V, Vcpu>, Self::Error>
    where
        V: VmArch,
        Vcpu: VcpuArch;

    fn configure_vcpu<V: Vm>(
        vm: &V, 
        hypervisor: &dyn HypervisorArch, 
        irq_chip: &mut dyn IrqChipArch, 
        vcpu: &mut dyn VcpuArch, 
        vcpu_init: &VcpuInitArch, 
        vcpu_id: usize, 
        num_cpus: usize, 
        has_bios: bool, 
        no_smt: bool, 
        host_cpu_topology: bool, 
        enable_pnp_data: bool, 
        itmt: bool
    ) -> Result<(), Self::Error>;

    fn register_pci_device<V: VmArch, Vcpu: VcpuArch>(
        linux: &mut RunnableLinuxVm<V, Vcpu>, 
        device: Box<dyn PciDevice>, 
        minijail: Option<Minijail>, 
        resources: &mut SystemAllocator
    ) -> Result<PciAddress, Self::Error>;

    fn debug_read_registers<T: VcpuArch>(
        vcpu: &T
    ) -> Result<GdbStubRegs, Self::Error>;

    fn debug_write_registers<T: VcpuArch>(
        vcpu: &T, 
        regs: &GdbStubRegs
    ) -> Result<(), Self::Error>;

    fn debug_read_memory<T: VcpuArch>(
        vcpu: &T, 
        guest_mem: &GuestMemory, 
        vaddr: GuestAddress, 
        len: usize
    ) -> Result<Vec<u8>, Self::Error>;

    fn debug_write_memory<T: VcpuArch>(
        vcpu: &T, 
        guest_mem: &GuestMemory, 
        vaddr: GuestAddress, 
        buf: &[u8]
    ) -> Result<(), Self::Error>;

    fn debug_enable_singlestep<T: VcpuArch>(vcpu: &T) -> Result<(), Self::Error>;

    fn debug_set_hw_breakpoints<T: VcpuArch>(
        vcpu: &T, 
        breakpoints: &[GuestAddress]
    ) -> Result<(), Self::Error>;
}

Trait which is implemented for each Linux Architecture in order to set up the memory, cpus, and system devices and to boot the kernel.

Associated Types

type Error: StdError

Required methods

fn guest_memory_layout(
components: &VmComponents
) -> Result<Vec<(GuestAddress, u64)>, Self::Error>

Returns a Vec of the valid memory addresses as pairs of address and length. These should be used to configure the GuestMemory structure for the platform.

Arguments

components - Parts used to determine the memory layout.

fn get_system_allocator_config<V: Vm>(vm: &V) -> SystemAllocatorConfig

Gets the configuration for a new SystemAllocator that fits the given Vm’s memory layout.

This is the per-architecture template for constructing the SystemAllocator. Platform agnostic modifications may be made to this configuration, but the final SystemAllocator will be at least as strict as this configuration.

Arguments

vm - The virtual machine to be used as a template for the SystemAllocator.

fn build_vm<V, Vcpu>(
components: VmComponents,
vm_evt_wrtube: &SendTube,
system_allocator: &mut SystemAllocator,
serial_parameters: &BTreeMap<(SerialHardware, u8), SerialParameters>,
serial_jail: Option<Minijail>,
battery: (&Option<BatteryType>, Option<Minijail>),
vm: V,
ramoops_region: Option<RamoopsRegion>,
devices: Vec<(Box<dyn BusDeviceObj>, Option<Minijail>)>,
irq_chip: &mut dyn IrqChipArch,
vcpu_ids: &mut Vec<usize>,
debugcon_jail: Option<Minijail>
) -> Result<RunnableLinuxVm<V, Vcpu>, Self::Error> where
V: VmArch,
Vcpu: VcpuArch,

Takes VmComponents and generates a RunnableLinuxVm.

Arguments

components - Parts to use to build the VM.

vm_evt_wrtube - Tube used by sub-devices to request that crosvm exit because guest wants to stop/shut down or requested reset.

system_allocator - Allocator created by this trait’s implementation of get_system_allocator_config.

serial_parameters - Definitions for how the serial devices should be configured.

serial_jail - Jail used for serial devices created here.

battery - Defines what battery device will be created.

vm - A VM implementation to build upon.

ramoops_region - Region allocated for ramoops.

devices - The devices to be built into the VM.

irq_chip - The IRQ chip implemention for the VM.

debugcon_jail - Jail used for debugcon devices created here.

fn configure_vcpu<V: Vm>(
vm: &V,
hypervisor: &dyn HypervisorArch,
irq_chip: &mut dyn IrqChipArch,
vcpu: &mut dyn VcpuArch,
vcpu_init: &VcpuInitArch,
vcpu_id: usize,
num_cpus: usize,
has_bios: bool,
no_smt: bool,
host_cpu_topology: bool,
enable_pnp_data: bool,
itmt: bool
) -> Result<(), Self::Error>

Configures the vcpu and should be called once per vcpu from the vcpu’s thread.

Arguments

vm - The virtual machine object.

hypervisor - The Hypervisor that created the vcpu.

irq_chip - The IrqChip associated with this vm.

vcpu - The VCPU object to configure.

vcpu_init - The data required to initialize VCPU registers and other state.

vcpu_id - The id of the given vcpu.

num_cpus - Number of virtual CPUs the guest will have.

has_bios - Whether the VmImage is a Bios image

no_smt - Wheter diabling SMT.

host_cpu_topology - whether enabling host cpu topology.

enable_pnp_data - whether enabling PnP statistics data.

itmt - whether enabling ITMT scheduler

fn register_pci_device<V: VmArch, Vcpu: VcpuArch>(
linux: &mut RunnableLinuxVm<V, Vcpu>,
device: Box<dyn PciDevice>,
minijail: Option<Minijail>,
resources: &mut SystemAllocator
) -> Result<PciAddress, Self::Error>

Configures and add a pci device into vm

fn debug_read_registers<T: VcpuArch>(
vcpu: &T
) -> Result<GdbStubRegs, Self::Error>

Reads vCPU’s registers.

fn debug_write_registers<T: VcpuArch>(
vcpu: &T,
regs: &GdbStubRegs
) -> Result<(), Self::Error>

Writes vCPU’s registers.

fn debug_read_memory<T: VcpuArch>(
vcpu: &T,
guest_mem: &GuestMemory,
vaddr: GuestAddress,
len: usize
) -> Result<Vec<u8>, Self::Error>

Reads bytes from the guest memory.

fn debug_write_memory<T: VcpuArch>(
vcpu: &T,
guest_mem: &GuestMemory,
vaddr: GuestAddress,
buf: &[u8]
) -> Result<(), Self::Error>

Writes bytes to the specified guest memory.

fn debug_enable_singlestep<T: VcpuArch>(vcpu: &T) -> Result<(), Self::Error>

Make the next vCPU’s run single-step.

fn debug_set_hw_breakpoints<T: VcpuArch>(
vcpu: &T,
breakpoints: &[GuestAddress]
) -> Result<(), Self::Error>

Set hardware breakpoints at the given addresses.