Crate arch

Re-exports

pub use serial::add_serial_devices;

pub use serial::get_serial_cmdline;

pub use serial::set_default_serial_parameters;

pub use serial::GetSerialCmdlineError;

pub use serial::SERIAL_ADDR;

Modules

android

fdt

This module writes Flattened Devicetree blobs as defined here: https://devicetree-specification.readthedocs.io/en/stable/flattened-format.html

pstore

serial

Structs

Pstore

RunnableLinuxVm

Holds the elements needed to run a Linux VM. Created by build_vm.

VirtioDeviceStub

The device and optional jail.

VmComponents

Holds the pieces needed to build a VM. Passed to build_vm in the LinuxArch trait below to create a RunnableLinuxVm.

Enums

DeviceRegistrationError

Errors for device manager.

LoadImageError

Errors for image loading.

VcpuAffinity

Mapping of guest VCPU threads to host CPU cores.

VmImage

Traits

LinuxArch

Trait which is implemented for each Linux Architecture in order to set up the memory, cpus, and system devices and to boot the kernel.

Functions

add_goldfish_battery

Adds goldfish battery return the platform needed resouces include its AML data, irq number

configure_pci_device

Config a PCI device for used by this vm.

generate_pci_root

Creates a root PCI device for use by this Vm.

generate_platform_bus

Creates a platform device for use by this Vm.

load_image

Load an image from a file into guest memory.

load_image_high

Load an image from a file into guest memory at the highest possible address.