Module arch::serial

Modules

sys

Enums

GetSerialCmdlineError

Constants

SERIAL_ADDR

Address for Serial ports in x86

Functions

add_serial_devices

Adds serial devices to the provided bus based on the serial parameters given.

get_serial_cmdline

Add serial options to the provided cmdline based on serial_parameters. serial_io_type should be “io” if the platform uses x86-style I/O ports for serial devices or “mmio” if the serial ports are memory mapped.

set_default_serial_parameters

Add the default serial parameters for serial ports that have not already been specified.

Type Definitions

GetSerialCmdlineResult